+++ packages/c-input/src/c-input-element.tsx
@@ -1,6 +1,7 @@
 import { computed, h, defineComponent, PropType } from "vue"
 import {
   chakra,
+  ComponentWithProps,
   HTMLChakraProps,
   SystemStyleObject,
   useStyles,
@@ -58,6 +59,7 @@ export const CInputLeftElement = defineComponent({
   name: "CInputLeftElement",
   setup(_, { attrs, slots }) {
     return () => (
+      // @ts-expect-error Untyped internal prop
       <CInputElement placement="left" __label="input__left-element" {...attrs}>
         {slots}
       </CInputElement>
@@ -74,6 +76,7 @@ export const CInputRightElement = defineComponent({
     return () => (
       <CInputElement
         placement="right"
+        // @ts-expect-error Untyped internal prop
         __label="input__right-element"
         {...attrs}
       >