+++ packages/c-modal/src/c-drawer.tsx
@@ -7,8 +7,6 @@ import {
   unref,
   withDirectives,
   watch,
-  ref,
-  Ref,
   watchEffect,
 } from 'vue'
 import {
@@ -65,9 +63,7 @@ const [
   useDrawerContext,
 ] = createContext<CDrawerContext>()
 
-export const CDrawer: ComponentWithProps<
-  DeepPartial<DrawerProps>
-> = defineComponent({
+export const CDrawer: ComponentWithProps<DrawerProps> = defineComponent({
   name: 'CDrawer',
   props: {
     ...modalProps,
@@ -116,9 +112,7 @@ export const CDrawer: ComponentWithProps<
 
 export interface DrawerContentProps extends HTMLChakraProps<'section'> {}
 
-export const CDrawerContent: ComponentWithProps<
-  DeepPartial<DrawerContentProps>
-> = defineComponent({
+export const CDrawerContent: ComponentWithProps<DrawerContentProps> = defineComponent({
   name: 'CDrawerContent',
   inheritAttrs: false,
   emits: ['click', 'mousedown', 'keydown'],
@@ -127,6 +121,7 @@ export const CDrawerContent: ComponentWithProps<
       dialogContainerProps: rawDialogContainerProps,
       dialogProps: rawDialogProps,
       modelValue,
+      blockScrollOnMount
     } = unref(useModalContext())
     const transitionId = useId('drawer-transition')
 
@@ -156,6 +151,27 @@ export const CDrawerContent: ComponentWithProps<
       outline: 0,
       ...styles.value.dialog,
     }))
+
+    // Scroll lock
+    watchEffect((onInvalidate) => {
+      if (!blockScrollOnMount!.value) return
+      if (modelValue.value !== true) return
+
+      let overflow = document.documentElement.style.overflow
+      let paddingRight = document.documentElement.style.paddingRight
+
+      let scrollbarWidth =
+        window.innerWidth - document.documentElement.clientWidth
+
+      document.documentElement.style.overflow = "hidden"
+      document.documentElement.style.paddingRight = `${scrollbarWidth}px`
+
+      onInvalidate(() => {
+        document.documentElement.style.overflow = overflow
+        document.documentElement.style.paddingRight = paddingRight
+        console.log("invalidating", document.documentElement.style.overflow)
+      })
+    })
 
     /** Handles exit transition */
     const leave = (done: VoidFunction) => {