+++ packages/c-modal/src/use-modal.ts
@@ -13,16 +13,19 @@ import {
   watchEffect,
 } from "vue"
 import { useIds } from "@chakra-ui/vue-composables"
-import { FocusLockProps, useFocusLock } from "@chakra-ui/c-focus-lock"
+import {
+  FocusLockProps,
+  useReturnFocusSelector,
+  useFocusTrap,
+} from "@chakra-ui/c-focus-lock"
 import {
   MaybeElementRef,
   useRef,
   getSelector,
   TemplateRef,
 } from "@chakra-ui/vue-utils"
-import { hideOthers, Undo } from "@chakra-ui/vue-a11y"
+import { hideOthers, Undo, useInertOthers } from "@chakra-ui/vue-a11y"
 import { focus, FocusableElement, isFunction } from "@chakra-ui/utils"
-import { useBodyScrollLock } from "@chakra-ui/c-scroll-lock"
 
 export interface UseModalOptions {
   /**
@@ -120,7 +123,7 @@ export function useModal(options: UseModalOptions) {
     return finalFocus
   })
 
-  const initialFocusElement = computed(() => {
+  const initialFocusElement = computed<HTMLElement>(() => {
     let initialFocus
     if (initialFocusRef?.value) {
       let resolvedInitialFocusRef: MaybeElementRef =
@@ -162,38 +165,13 @@ export function useModal(options: UseModalOptions) {
 
   const hasHeader = ref(false)
   const hasBody = ref(false)
-
-  /** Initialize focus lock */
-  const { lock, deactivate } = useFocusLock({
-    escapeDeactivates: false,
-    clickOutsideDeactivates: false,
-    allowOutsideClick: true,
-    returnFocusOnDeactivate: true,
-    delayInitialFocus: true,
-    onActivate() {
-      if (initialFocusElement.value) {
-        setTimeout(() => {
-          focus(initialFocusElement.value)
-        })
-      }
-    },
-    onDeactivate() {
-      /**
-       * There appears to be a bug in which
-       * the DOM refreshes and elements are modified
-       * in a way that completely replaces elements in the DOM
-       * such that the targeted nodes are not found in the
-       * browser even though we have them in JavaScript.
-       *
-       * At the time of writing this composable, I am
-       * unable to ascertain where this issue came from.
-       *
-       * However, as a failsafe, ew allow the `useModal()`
-       * hook to always track the last focused element
-       * before it was activated using the `useReturnFocusSelector`
-       *
-       * @see useReturnFocusSelector Function
-       */
+  /**
+   * This watcher is being used to track
+   * the element refs for the dialog container
+   * element.
+   */
+  watch(dialogRefEl, (newVal) => {
+    if (!newVal) {
       setTimeout(() => {
         const lastfocusedNode = document.querySelector(
           lastFocusedSelector.value as string
@@ -205,48 +183,52 @@ export function useModal(options: UseModalOptions) {
           focus(lastfocusedNode as HTMLElement)
         }
       }, 100)
-    },
-    immediate: true,
-  })
-
-  const { scrollLockRef } = useBodyScrollLock(
-    blockScrollOnMount?.value ? modelValue : ref(false)
-  )
-
-  /**
-   * This watcher is being used to track
-   * the element refs for the dialog container
-   * element.
-   *
-   * When the ref is bound, we activate
-   * the focus lock and body scroll lock refs.
-   */
-  watch(dialogRefEl, (newVal) => {
-    if (newVal) {
-      lock(newVal)
-      scrollLockRef(newVal)
-    } else {
-      deactivate()
     }
   })
+
+  const containers = ref<Set<HTMLElement>>(new Set())
+
+  watchEffect(
+    (onInvalidate) => {
+      let el: HTMLElement
+      if (dialogRefEl.value) {
+        el = dialogRefEl.value
+        containers.value.add(el)
+      }
+
+      onInvalidate(() => {
+        containers.value.delete(el)
+      })
+    },
+    { flush: "post" }
+  )
+
+  useFocusTrap(
+    containers,
+    ref(true),
+    computed(() => ({
+      initialFocus: initialFocusElement.value,
+    }))
+  )
 
   /**
    * Dialog props
    */
   const dialogProps = computed<(context: any) => MergedVNodeProps>(
-    () => ({ emit }) => ({
-      role: "dialog",
-      ref: dialogRef,
-      id: dialogId.value,
-      tabIndex: -1,
-      "aria-modal": true,
-      "aria-labelledby": hasHeader.value ? headerId.value : null,
-      "aria-describedby": hasBody.value ? bodyId.value : null,
-      onClick(event: MouseEvent) {
-        event.stopPropagation()
-        emit("click", event)
-      },
-    })
+    () =>
+      ({ emit }) => ({
+        role: "dialog",
+        ref: dialogRef,
+        id: dialogId.value,
+        tabIndex: -1,
+        "aria-modal": true,
+        "aria-labelledby": hasHeader.value ? headerId.value : null,
+        "aria-describedby": hasBody.value ? bodyId.value : null,
+        onClick(event: MouseEvent) {
+          event.stopPropagation()
+          emit("click", event)
+        },
+      })
   )
 
   const handleOverlayClick = (event: MouseEvent) => {
@@ -273,22 +255,23 @@ export function useModal(options: UseModalOptions) {
 
   /** Dialog container props */
   const dialogContainerProps = computed<(context: any) => MergedVNodeProps>(
-    () => ({ emit }) => ({
-      ref: overlayRef as any,
-      onClick: (event: MouseEvent) => {
-        instance?.emit("update:modelValue", !modelValue.value)
-        instance?.emit("closeModal")
-        handleOverlayClick(event)
-      },
-      onKeydown: (event: KeyboardEvent) => {
-        emit("keydown", event)
-        onKeyDown(event)
-      },
-      onMousedown: (event: MouseEvent) => {
-        mouseDownTarget.value = event.target
-        emit("mousedown", event)
-      },
-    })
+    () =>
+      ({ emit }) => ({
+        ref: overlayRef as any,
+        onClick: (event: MouseEvent) => {
+          instance?.emit("update:modelValue", !modelValue.value)
+          instance?.emit("closeModal")
+          handleOverlayClick(event)
+        },
+        onKeydown: (event: KeyboardEvent) => {
+          emit("keydown", event)
+          onKeyDown(event)
+        },
+        onMousedown: (event: MouseEvent) => {
+          mouseDownTarget.value = event.target
+          emit("mousedown", event)
+        },
+      })
   )
   /**
    * `aria-hidden` attributes handling
@@ -297,7 +280,8 @@ export function useModal(options: UseModalOptions) {
   const shouldHide = computed(
     () => (modelValue.value && useInert?.value) || false
   )
-  useAriaHidden(dialogRefEl, shouldHide)
+  // useAriaHidden(dialogRefEl, shouldHide)
+  useInertOthers(dialogRefEl, shouldHide)
 
   return {
     modelValue,
@@ -346,31 +330,4 @@ export function useAriaHidden(
       flush: "post",
     }
   )
-}
-
-/** Tracks last focused element selector before Modal/dialog is opened */
-export function useReturnFocusSelector(shouldTrack: Ref<boolean>) {
-  const lastFocused = ref<EventTarget | null>(null)
-  const lastFocusedSelector = ref<string | undefined>()
-
-  const trackFocus = (event: Event) => {
-    if (!shouldTrack.value) {
-      lastFocusedSelector.value = getSelector(event.target as HTMLElement)
-    }
-  }
-
-  onBeforeMount(() => {
-    document.addEventListener("focusin", trackFocus)
-  })
-
-  onBeforeUnmount(() => {
-    document.removeEventListener("focusin", trackFocus)
-    lastFocused.value = null
-    lastFocusedSelector.value = undefined
-  })
-
-  return {
-    lastFocused,
-    lastFocusedSelector,
-  }
 }