+++ packages/utils/src/configurable.ts
@@ -1,0 +1,3 @@
+import { canUseDOM } from "@chakra-ui/utils"
+
+export const defaultWindow = canUseDOM() ? window : null