+++ packages/vue-composables/src/index.ts
@@ -1,1 +1,5 @@
 export * from "./use-id"
+export * from "./use-event-listener"
+export * from "./use-window-event"
+export * from "./use-element-stack"
+export * from "./use-clipboard"