+++ website/src/docs-theme/layout/LayoutMdx.vue
@@ -3,9 +3,7 @@
     <template #sidebar>
       <app-sidebar :routes="routes" />
     </template>
-    <template>
-      <slot name="default" />
-    </template>
+    <slot />
     <template #pagination>
       <app-pagination
         :previous="routeContext.prevRoute"
@@ -16,16 +14,16 @@
 </template>
 
 <script lang="ts">
-import { defineComponent, PropType } from 'vue'
-import { getHeadings } from '@/docs-theme/utils/get-headings'
-import { getRoutes } from '@/docs-theme/utils/get-routes'
-import { useRoute } from 'vue-router'
-import { getRouteContext } from '@/docs-theme/utils/get-route-context'
+import { defineComponent, PropType } from "vue"
+import { getHeadings } from "@/docs-theme/utils/get-headings"
+import { getRoutes } from "@/docs-theme/utils/get-routes"
+import { useRoute } from "vue-router"
+import { getRouteContext } from "@/docs-theme/utils/get-route-context"
 
 import {
   findRouteByPath,
   removeFromLast,
-} from '@/docs-theme/utils/find-route-by-path'
+} from "@/docs-theme/utils/find-route-by-path"
 
 export default defineComponent({
   props: {
@@ -43,7 +41,7 @@ export default defineComponent({
     const headings = getHeadings(slots)
 
     const route = findRouteByPath(
-      removeFromLast(props.frontmatter.slug, '#'),
+      removeFromLast(props.frontmatter.slug, "#"),
       routes
     )