+++ packages/c-form-control/CHANGELOG.md
@@ -1,16 +1,38 @@
 # @chakra-ui/c-form-control
 
-## 0.0.0-alpha.1
+## 0.0.0-alpha.3
+
+### Patch Changes
+
+- lz "chore: upgrade to latest version. Fixes build error"
+
+- Updated dependencies []:
+  - @chakra-ui/c-icon@1.0.0-alpha.8
+  - @chakra-ui/vue-system@0.1.0-alpha.8
+  - @chakra-ui/vue-utils@0.1.0-alpha.8
+  - @chakra-ui/vue-composables@0.1.0-alpha.7
+
+## 0.0.0-alpha.2
+
 ### Minor Changes
 
+- chore: parity release
 
+### Patch Changes
+
+- Updated dependencies []:
+  - @chakra-ui/c-icon@1.0.0-alpha.7
+  - @chakra-ui/vue-system@0.1.0-alpha.7
+  - @chakra-ui/vue-utils@0.1.0-alpha.7
+  - @chakra-ui/vue-composables@0.1.0-alpha.6
+
+## 0.0.0-alpha.1
+
+### Minor Changes
 
 - [#108](https://github.com/chakra-ui/chakra-ui-vue-next/pull/108) [`a7df008`](https://github.com/chakra-ui/chakra-ui-vue-next/commit/a7df008af57158fbd9240645d4177cf862ca764d) Thanks [@codebender828](https://github.com/codebender828)! - Create form control and input components
 
-
-
-- [#91](https://github.com/chakra-ui/chakra-ui-vue-next/pull/91) [`64ec694`](https://github.com/chakra-ui/chakra-ui-vue-next/commit/64ec694a67bf0932ad8d7569cf4347cd1da48513) Thanks [@codebender828](https://github.com/codebender828)! - Create `CFormControl` component
-
+* [#91](https://github.com/chakra-ui/chakra-ui-vue-next/pull/91) [`64ec694`](https://github.com/chakra-ui/chakra-ui-vue-next/commit/64ec694a67bf0932ad8d7569cf4347cd1da48513) Thanks [@codebender828](https://github.com/codebender828)! - Create `CFormControl` component
 
 ### Patch Changes