+++ packages/nuxt/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @chakra-ui/nuxt-next
 
+## 0.1.0-alpha.5
+
+### Patch Changes
+
+- lz "chore: upgrade to latest version. Fixes build error"
+
+## 0.1.0-alpha.4
+
+### Minor Changes
+
+- chore: parity release
+
 ## 0.1.0-alpha.3
 
 ### Minor Changes