+++ packages/theme-tools/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @chakra-ui/vue-theme-tools
 
+## 0.1.0-alpha.8
+
+### Patch Changes
+
+- lz "chore: upgrade to latest version. Fixes build error"
+
+- Updated dependencies []:
+  - @chakra-ui/vue-system@0.1.0-alpha.8
+
+## 0.1.0-alpha.7
+
+### Minor Changes
+
+- chore: parity release
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @chakra-ui/vue-system@0.1.0-alpha.7
+
 ## 0.1.0-alpha.6
 
 ### Minor Changes